Splits renderers into separate subcomponent
- Creates subcomponent `Hal\Renderer`
- Creates interface `Hal\Renderer\Renderer`
- Creates class `Hal\Renderer\JsonRenderer`
- Creates class `Hal\Renderer\XmlRenderer`
- Modifies `Hal\HalResponseFactory` to create a `JsonRenderer` and
  `XmlRenderer` during instantiation, and delegate to those when
  creating a response.
