+++ src/HalResponseFactory.php
@@ -3,8 +3,6 @@
 namespace Hal;
 
 use Closure;
-use DOMDocument;
-use DOMNode;
 use Negotiation\Negotiator;
 use Psr\Http\Message\ResponseInterface;
 use Psr\Http\Message\ServerRequestInterface;
@@ -33,14 +31,10 @@ class HalResponseFactory
         'application/*+xml',
     ];
 
-    /**
-     * @var int
-     */
-    private $jsonFlags;
+    /** @var Renderer\JsonRenderer */
+    private $jsonRenderer;
 
-    /**
-     * @var ResponseInterface
-     */
+    /** @var ResponseInterface */
     private $responsePrototype;
 
     /**
@@ -51,12 +45,16 @@ class HalResponseFactory
      */
     private $streamFactory;
 
+    /** @var Renderer\XmlRenderer */
+    private $xmlRenderer;
+
     public function __construct(
         int $jsonFlags = self::DEFAULT_JSON_FLAGS,
         ResponseInterface $responsePrototype = null,
         callable $streamFactory = null
     ) {
-        $this->jsonFlags = $jsonFlags;
+        $this->jsonRenderer = new Renderer\JsonRenderer($jsonFlags);
+        $this->xmlRenderer = new Renderer\XmlRenderer();
         $this->responsePrototype = $responsePrototype ?: new Response();
         $this->streamFactory = $streamFactory ?: Closure::fromCallable([$this, 'generateStream']);
     }
@@ -70,18 +68,23 @@ class HalResponseFactory
         $matchedType = (new Negotiator())->getBest($accept, self::NEGOTIATION_PRIORITIES);
 
         switch (true) {
+            case ($matchedType && strstr($matchedType->getValue(), 'json')):
+                $renderer = $this->jsonRenderer;
+                $mediaType = $mediaType . '+json';
+                break;
             case (! $matchedType):
-                $generator = Closure::fromCallable([$this, 'generateXmlResponse']);
-                break;
-            case (strstr($matchedType->getValue(), 'json')):
-                $generator = Closure::fromCallable([$this, 'generateJsonResponse']);
-                break;
+                // fall-through
             default:
-                $generator = Closure::fromCallable([$this, 'generateXmlResponse']);
+                $renderer = $this->xmlRenderer;
+                $mediaType = $mediaType . '+xml';
                 break;
         }
 
-        return $generator($resource, $mediaType);
+        $body = ($this->streamFactory)();
+        $body->write($renderer->render($resource));
+        return $this->responsePrototype
+            ->withBody($body)
+            ->withHeader('Content-Type', $mediaType);
     }
 
     /**
@@ -105,152 +108,5 @@ class HalResponseFactory
     private function generateStream() : Stream
     {
         return new Stream('php://temp', 'wb+');
-    }
-
-    private function generateJsonResponse(HalResource $resource, string $mediaType) : ResponseInterface
-    {
-        $body = ($this->streamFactory)();
-        $body->write(json_encode($resource, $this->jsonFlags));
-        return $this->responsePrototype
-            ->withBody($body)
-            ->withHeader('Content-Type', $mediaType . '+json');
-    }
-
-    private function generateXmlResponse(HalResource $resource, string $mediaType) : ResponseInterface
-    {
-        $body = ($this->streamFactory)();
-        $body->write($this->createXmlPayload($resource));
-        return $this->responsePrototype
-            ->withBody($body)
-            ->withHeader('Content-Type', $mediaType . '+xml');
-    }
-
-    private function createXmlPayload(HalResource $resource) : string
-    {
-        $dom = new DOMDocument('1.0', 'UTF-8');
-        $dom->formatOutput = true;
-        $dom->appendChild($this->createResourceNode($dom, $resource->toArray()));
-        return trim($dom->saveXML());
-    }
-
-    private function createResourceNode(DOMDocument $doc, array $resource, string $resourceRel = 'self') : DOMNode
-    {
-        // Normalize resource
-        $resource['_links']    = $resource['_links'] ?? [];
-        $resource['_embedded'] = $resource['_embedded'] ?? [];
-
-        $node = $doc->createElement('resource');
-
-        // Self-relational link attributes, if present and singular
-        if (isset($resource['_links']['self']['href'])) {
-            $node->setAttribute('rel', $resourceRel);
-            $node->setAttribute('href', $resource['_links']['self']['href']);
-            foreach ($resource['_links']['self'] as $attribute => $value) {
-                if ($attribute === 'href') {
-                    continue;
-                }
-                $node->setAttribute($attribute, $value);
-            }
-            unset($resource['_links']['self']);
-        }
-
-        foreach ($resource['_links'] as $rel => $linkData) {
-            if ($this->isAssocArray($linkData)) {
-                $node->appendChild($this->createLinkNode($doc, $rel, $linkData));
-                continue;
-            }
-
-            foreach ($linkData as $linkDatum) {
-                $node->appendChild($this->createLinkNode($doc, $rel, $linkDatum));
-            }
-        }
-        unset($resource['_links']);
-
-        foreach ($resource['_embedded'] as $rel => $childData) {
-            if ($this->isAssocArray($childData)) {
-                $node->appendChild($this->createResourceNode($doc, $childData, $rel));
-                continue;
-            }
-
-            foreach ($childData as $childDatum) {
-                $node->appendChild($this->createResourceNode($doc, $childDatum, $rel));
-            }
-        }
-        unset($resource['_embedded']);
-
-        return $this->createNodeTree($doc, $node, $resource);
-    }
-
-    private function createLinkNode(DOMDocument $doc, string $rel, array $data)
-    {
-        $link = $doc->createElement('link');
-        $link->setAttribute('rel', $rel);
-        foreach ($data as $key => $value) {
-            $value = $this->normalizeConstantValue($value);
-            $link->setAttribute($key, $value);
-        }
-        return $link;
-    }
-
-    /**
-     * Convert true, false, and null to appropriate strings.
-     *
-     * In all other cases, return the value as-is.
-     *
-     * @param mixed $value
-     * @return string|mixed
-     */
-    private function normalizeConstantValue($value)
-    {
-        $value = $value === true ? 'true' : $value;
-        $value = $value === false ? 'false' : $value;
-        $value = $value === null ? '' : $value;
-        return $value;
-    }
-
-    private function isAssocArray(array $value) : bool
-    {
-        return array_values($value) !== $value;
-    }
-
-    /**
-     * @return DOMNode|DOMNode[]
-     */
-    private function createResourceElement(DOMDocument $doc, string $name, $data)
-    {
-        if (is_scalar($data)) {
-            $data = $this->normalizeConstantValue($data);
-            return $doc->createElement($name, $data);
-        }
-
-        if (! is_array($data)) {
-            throw Exception\InvalidResourceValueException::fromValue($data);
-        }
-
-        if ($this->isAssocArray($data)) {
-            return $this->createNodeTree($doc, $doc->createElement($name), $data);
-        }
-
-        $elements = [];
-        foreach ($value as $child) {
-            $elements[] = $this->createResourceElement($doc, $name, $child);
-        }
-        return $elements;
-    }
-
-    private function createNodeTree(DOMDocument $doc, DOMNode $node, array $data) : DOMNode
-    {
-        foreach ($data as $key => $value) {
-            $element = $this->createResourceElement($doc, $key, $value);
-            if (! is_array($element)) {
-                $node->appendChild($element);
-                continue;
-            }
-            foreach ($element as $child) {
-                $node->appendChild($child);
-            }
-        }
-
-        return $node;
     }
 }
