+++ src/Renderer/Renderer.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Hal\Renderer;
+
+use Hal\HalResource;
+
+interface Renderer
+{
+    public function render(HalResource $resource) : string;
+}
